pyproject.toml
[tool.poetry]
name = "aigent"
version = "0.1.0"
description = "AIGENT: Intelligent Document Processing and AI Agent Toolkit"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "aigent"}]

[tool.poetry.dependencies]
python = "^3.8"
thinc = "^8.1.0"
pydantic = "^1.10.0"
typer = "^0.9.0"
fastapi = "^0.100.0"
gradio = "^3.50.2"
jupyterlab = "^4.0.0"
spacy = {git = "https://github.com/explosion/spaCy.git", rev = "v3.7.2"}
ipykernel = "^6.29.5"
pdfplumber = "^0.11.2"
pypdf2 = "^3.0.1"
iprogress = "^0.4"
pyqt5 = "^5.15.11"
sqlalchemy = "^2.0.31"
pytesseract = "^0.3.10"
pillow = "^10.4.0"
opencv-python = "^4.10.0.84"
openai = "^1.37.0"
anthropic = "^0.31.2"
google-cloud-vision = "^3.4.0"
google-auth-oauthlib = "^1.0.0"
opencv-python-headless = "^4.10.0.84"
python-docx = "^0.8.11"
python-dotenv = "^1.0.1"
